SRV-2819: Prevent FileNotFoundException when moving directories.
SRV-2819 (Overwrite option to keep or change history when moving/copying workflows)

+++ org.knime.core.util/src/main/java/org/knime/core/util/PathUtils.java
@@ -435,7 +435,7 @@ public final class PathUtils {
      * @throws IOException if an I/O error occurs while copying
      */
     public static void moveDirectory(final Path source, final Path destination) throws IOException {
-        if (!Files.isDirectory(destination)
+        if (!Files.isDirectory(destination) && destination.getParent().toFile().exists() && source.toFile().exists()
             && Files.getFileStore(source).equals(Files.getFileStore(destination.getParent()))) {
             Files.move(source, destination);
         } else {
